fix remaining integration tests
