+++ pynamodb/settings.py
@@ -14,7 +14,6 @@ default_settings_dict = {
     'region': 'us-east-1',
     'max_pool_connections': 10,
     'extra_headers': None,
-    'allow_rate_limited_scan_without_consumed_capacity': False,
 }
 
 OVERRIDE_SETTINGS_PATH = getenv('PYNAMODB_CONFIG', '/etc/pynamodb/global_default_settings.py')
