--- pynamodb/tests/integration/config.py
@@ -1,5 +1,0 @@
-"""
-Integration test settings
-"""
-
-DYNAMODB_HOST = 'http://localhost:8000'
